pyproject.toml
[project]
name = "genai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.3",
    "arxiv>=2.4.1",
    "chromadb>=1.5.5",
    "duckduckgo-search>=8.1.1",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.1",
    "langchain>=1.2.13",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.20",
    "langchain-google-genai>=4.2.1",
    "langchain-mistralai>=1.1.2",
    "langchain-ollama>=1.0.1",
    "numpy>=2.4.3",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "streamlit>=1.55.0",
    "tiktoken>=0.12.0",
    "tqdm>=4.67.3",
    "unstructured>=0.21.5",
    "uvicorn>=0.42.0",
    "wikipedia>=1.4.0",
]
